Refactored code to fix some more code duplication
