+++ src/ClassFramework.Pipelines/Builder/Components/AddFluentMethodsForNonCollectionPropertiesComponent.cs
@@ -49,23 +49,9 @@ public class AddFluentMethodsForNonCollectionPropertiesComponent : IPipelineComp
                 .WithReturnTypeName(context.Context.IsBuilderForAbstractEntity
                       ? $"TBuilder{context.Context.SourceModel.GetGenericTypeArgumentsString()}"
                       : $"{results.First(x => x.Name == "Namespace").Result.Value.AppendWhenNotNullOrEmpty(".")}{results.First(x => x.Name == "BuilderName").Result.Value}{context.Context.SourceModel.GetGenericTypeArgumentsString()}")
-                .AddParameters
-                (
-                    new ParameterBuilder()
-                        .WithName(property.Name.ToPascalCase(context.Context.FormatProvider.ToCultureInfo()))
-                        .WithTypeName(results.First(x => x.Name == "TypeName").Result.Value!)
-                        .SetTypeContainerPropertiesFrom(property)
-                        .WithDefaultValue(context.Context.GetMappingMetadata(property.TypeName).GetValue<object?>(MetadataNames.CustomBuilderWithDefaultPropertyValue, () => null))
-                );
+                .AddParameters(context.Context.CreateParameterForBuilder(property, results));
 
-            if (context.Context.Settings.AddNullChecks)
-            {
-                var nullCheckStatement = results.First(x => x.Name == "ArgumentNullCheck").Result.Value!;
-                if (!string.IsNullOrEmpty(nullCheckStatement))
-                {
-                    builder.AddStringCodeStatements(nullCheckStatement);
-                }
-            }
+            context.Context.AddNullChecks(builder, results);
 
             builder.AddStringCodeStatements
             (
